fix pg_alter_subscription_refresh_publication() on pg14
In Postgres 14 ALTER SUBSCRIPTION ... REFRESH cannot be executed from
a function anymore. To work around that we do this over a local dblink
connection.

It's not very elegant, but we already use dblink, so at least this
doesn't introduce a new dependency. For the sake of consistency we
do it the same way for older versions of Postgres as well.

Fixes #25.