Provide an aiven_extras.truncate_freespacemap function
When a bug involving FSM corruption, we currently have no easy way of fixing it.
We can either VACUUM FULL the relation, which will rewrite everything including the FSM. Or we can stop the service, delete the FSM manually, and restart it.

This function is copy-pasted from pg_truncate_visibility_map, and allows to truncate an FSM online.